+++ src/main/java/com/develhope/spring/configurations/SpringMVCConfiguration.java
@@ -1,8 +1,8 @@
 package com.develhope.spring.configurations;
 
-import com.develhope.spring.interceptors.AdminInterceptor;
-import com.develhope.spring.interceptors.ClientInterceptor;
-import com.develhope.spring.interceptors.SellerInterceptor;
+import com.develhope.spring.admin.AdminInterceptor;
+import com.develhope.spring.client.ClientInterceptor;
+import com.develhope.spring.seller.SellerInterceptor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
