--- src/main/java/com/develhope/spring/interceptors/UserInterceptor.java
@@ -1,4 +1,0 @@
-package com.develhope.spring.interceptors;
-
-public class UserInterceptor {
-}
