+++ src/main/java/com/develhope/spring/user/UserInterceptor.java
@@ -1,0 +1,4 @@
+package com.develhope.spring.user;
+
+public class UserInterceptor {
+}
